[WIP] server-side pseudo-form view
For complex objects which rely on onchange & other to reach a proper
state, creating model tests can be a PITA as onchanges either aren't
applied or are called one by one in an ad-hoc manner, which may not
correspond to how the object works (or is supposed to work).

This is a work in progress on an object formalizing that flow, and
allowing a "more proper" way to create and alter records server-side
when that's desirable.