[FIX] web_calendar: prevent crash when the color field is undefined
Add a check for the presence of color field. (only applies when "useContacts" is not set)

Before this fix, if the view xml does not set "useContacts", the calendar view fallback on the color attribute (not required) which may be non-existent on the model and on the view.
web_calendar.js:799 is using "if (color_field.type == "selection")" so its throwing this error when color is not defined:
Uncaught TypeError: Cannot read property 'type' of undefined