[FIX] account: try to improve payment info layout
The PaymentInfo widget uses a bunch of hardcoded CSS rules, which
sometimes makes the amounts very hard to read.
For example in the following screenshot in saas-15, the amount is
supposed to be 38.72 EUR, but that's hard to see:

![payment_layout](https://cloud.githubusercontent.com/assets/2039870/25992496/f7e12cf6-3706-11e7-9791-9337e06c7523.png)

Forcing `whitespace: nowrap` on the amount cell in Enterprise is done
via the `o_form_field_number` class. Replacing the `margin-right` inside
the TD with a `padding-right` on the TD also helps separating the
name of the line from the amount.

This whole widget should probably be cleaned up.

Done in 10.0 to avoid disrupting 9.0 as long as the problem is not too
common.