[MOV] website, web_routing: move translation support in web_routing
Code managing language in URLs and redirection is moved from website
to web_routing. Override of _dispatch in website now consists in adding
website to the request.

url_for and is_multilang_url tools methods are also moved from website
to web_routing.ir_http as those are required for the moved code.