[FIX] web_kanban: allow column creation on custom model
This is a back-port of the commit https://github.com/odoo/odoo/commit/1a2c05e98bba7694831c82b7be8bb81a9c46535c
that had been applied in saas-16.

As this has been considered as a bug, it is fixed in previous versions.
No need to forward-port it after saas-15 then.