[REF] models: automatic subqueries
Use a `LazyQuery` object to represent a sequence of ids obtained from an SQL
query.  The query is executed automatically whenever the sequence is accessed.
The method `search` behaves as before, but `_search` has been modified to
return a `LazyQuery` by default.

When using `LazyQuery` objects in domain conditions, the SQL query generator
uses the wrapped query instead of its result, which is how subqueries are
generated.  The search on sequence of relational fields now uses `_search`
instead of `search` to perform sub-searches and enable sub-selects when
possible.

This reduces the number of queries by around 10%, and helps avoiding
pathological SQL queries containing thousands of values injected from former
queries.