[FIX] purchase: correctly link purchase order and invoice
When creating a invoice from the purchase order, the later one wasn't not
correctly link anymore (and the invoice couldn't be found through the stat button.

The link is actually done with an onchange at the invoice creation.

The problem here was the readonly attribute set on the fields description but
was not overriden in the view. The onchange was taken into account in previous
versions but with the new views, the correct behaviour has been implemented: if
a field is readonly, the onchange won't write on it.