[IMP] project: optimize subtasks count
The process to show one thousand records in task
kanban view takes more than 8 minutes.

Using a search with `child_of` can be expensive in
terms of time, when a huge number of task/sub-tasks
exists in database.

The ORM propose the `_parent_store` to optimize
the usage of `child_of` operator.

Closes #18196