[FIX] stock : update quantity on hand in sub locations
When updating the quantity on hand in a location without quant but
where sub location has quants for the product. It will increase the
available quantity in the sub location instead of the defined location.

This happens because _gather method in quant search on child_of location
in any case.

This commit change the domain by using '=' in case of argument strict=True
or 'child_of' if it is set to False. It also remove the argument strict on
_update_avaible_quantity since there is no logic to use it with strict=false