[ADD] http_routing: advanced routing for web module
This module will contain website routing capabilities that are not
necessarily linked to a website. This contains for example code related
to language management in URLs or slugify support. As we don't want to
put that code directly in base in order to keep the core simple or in
web to limit it to web client only http_routing module will be used to
hold those features.

This commit only adds a void module, future commits will gradually add
the features.