[FIX] web: better context handling for sub views
When calling the route 'call_button', we need to add magic keys in the
context
(active_id, active_ids, active_model).  This was automatically done by
the view manager in the do_execute_action method.  However, in some
cases, it was wrong.  Here is such an example:

- open a manufacturing order in form view (manufacturing app)
- click on edit
- click on one of the line of the one2many, and click on the green icon
  to edit a product
- click on the update product quantity button on top of it

when sub modal => a call button may need to be called with the active
keys from the sub modal and not from the main view.  with this commit,
we add the keys properly...

need to finish this...