[FIX] stock : check backorder compare apple and orange.
The function taht check if we need a back oder or not compare
the initial demand and the quantity processed on the picking.

When unit of measure are activated the initial demand (product_qty)
use the uom of the product and the quantity processed on the move
(quantity_done) has the uom of the move.

This commit use product_uom_qty instead of product_qty. Product_uom_qty
represents the initial demand transformed with the uom of the move.
It also add a small test with a back order that ensure the quantity
are correct.