[FIX] account: setup bar: multi company in chart of accounts step
It was possible to edit the opening debit and credit of a different company than the one being configured. If no opening move had yet been computed, it triggered an error. Editing accounts from other companies from this configuration step is now impossible.

[IMP] account: setup bar: usuability of "bank accounts" step

Hide some advanced fields, to make the initial configuration easier.

[FIX] account: setup bar: layout correction

A margin was applied to the setup bar even when no data was displayed in the dashboard, which caused it to shift, in conjunction with o_kanban_no_content class. Now, this margin is only applied when there are data to display.